requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
numpy
scikit-learn
python-multipart
python-dotenv
pytest
statsmodels
pdf2image
Pillow
matplotlib
google-generativeai
pymupdf
pymysql
scikit-learn
sdv